pycaffe/stitchpyramid related makefile tweaks; add missing (PyArrayObject*) casts (to fix compile errors for some build envs).
note also that there is a new SHARED_LDFLAGS makefile var that can be set to include -Wl,--no-undefined (for gcc) to avoid accidentally linking an .so missing some of its dependencies. but, since people use other compilers, it's not enabled by default, and only the stitch library build line uses the macro at this point.

mwm